Selection 수정
무한 루프에 빠지는 경우에 시간이 오래 걸리는 경우가 발생할 수 있어 swap하여 중복하지 않는 짝을 만들도록 수정

+++ FBNDP-MLTN/Population/Population.h
@@ -8,7 +8,7 @@ using namespace std;
 class Population
 {
 public:
-	Population(int MemberNum);
+	Population(uint64_t MemberNum);
 	uint64_t GetCurrentChromosomeNum() { return ChromosomeArray.size(); }
 	Chromosome GetChromosome(uint64_t ChromosomeNum) { return ChromosomeArray.at(ChromosomeNum); }
 	vector<NodeData> GetChromosomeData(uint64_t ChromosomeNum)	/*ChromosomeNum번째의 ChromosomeData를 얻음*/
